requirements.txt
python-telegram-bot[job-queue]==20.3
aiohttp>=3.8.6
requests>=2.31.0